Moving motion value animation function
