+++ packages/framer-motion/src/animation/interfaces/motion-value.ts
@@ -1,17 +1,17 @@
-import { warning } from "../utils/errors"
-import { ResolvedValueTarget, Transition } from "../types"
-import { secondsToMilliseconds } from "../utils/time-conversion"
-import { instantAnimationState } from "../utils/use-instant-transition-state"
-import type { MotionValue, StartAnimation } from "../value"
-import { createAcceleratedAnimation } from "./animators/waapi/create-accelerated-animation"
-import { createInstantAnimation } from "./animators/instant"
-import { AnimationOptions } from "./types"
-import { getDefaultTransition } from "./utils/default-transitions"
-import { isAnimatable } from "./utils/is-animatable"
-import { getKeyframes } from "./utils/keyframes"
-import { getValueTransition, isTransitionDefined } from "./utils/transitions"
-import { animateValue } from "./animators/js"
-import { AnimationPlaybackControls } from "./types"
+import { warning } from "../../utils/errors"
+import { ResolvedValueTarget, Transition } from "../../types"
+import { secondsToMilliseconds } from "../../utils/time-conversion"
+import { instantAnimationState } from "../../utils/use-instant-transition-state"
+import type { MotionValue, StartAnimation } from "../../value"
+import { createAcceleratedAnimation } from "../animators/waapi/create-accelerated-animation"
+import { createInstantAnimation } from "../animators/instant"
+import { AnimationOptions } from "../types"
+import { getDefaultTransition } from "../utils/default-transitions"
+import { isAnimatable } from "../utils/is-animatable"
+import { getKeyframes } from "../utils/keyframes"
+import { getValueTransition, isTransitionDefined } from "../utils/transitions"
+import { animateValue } from "../animators/js"
+import { AnimationPlaybackControls } from "../types"
 
 export const createMotionValueAnimation = (
     valueName: string,
